+++ src/components/dashboard/SimpleWidget.tsx
