+++ src/components/index.ts
@@ -1,3 +1,4 @@
-export { Sidebar } from './Sidebar';
+export { Sidebar } from './sidebar/Sidebar';
+export { SidebarMenuItem } from './sidebar/SidebarMenuItem';
 
-export { SimpleWidget } from './SimpleWidget';
+export { SimpleWidget } from './dashboard/SimpleWidget';
